microservices/movie-service/requirements.txt
fastapi
sqlalchemy
databases[postgres]
psycopg2
asyncpg
fastapi-cache2[redis]